requirements.txt
google-ai-generativelanguage
instagrapi
langchain
langchain-core
langchain-google-genai
matplotlib
numpy
pandas
pillow
requests
selenium
streamlit
streamlit-dash
streamlit-js-eval
tqdm
webdriver-manager
